+++ mono_repo/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0
+
+* **BREAKING**
+  The package configuration file is now `mono_pkg.yaml`. If a legacy config file
+  – `.mono_repo.yml` – is found, the command is canceled and a warning is printed
+  telling the user to rename the file.
+
 ## 0.3.3
 
 * Support adding custom cache directories in each project.